Add some more useful page related functions
